Use caching ConsoleStream for both iOS and Android
Expands on #56713 by moving the caching implementation to a separate internal class leaving only the interop calls in ConsolePal.iOS and ConsolePal.Android

Fixes https://github.com/dotnet/runtime/issues/57304
